convexpire in its own .c file
